add user crud

+++ app/Models/User.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Cashier\Billable;
@@ -11,7 +12,7 @@ use Laravel\Cashier\Billable;
 class User extends Authenticatable implements MustVerifyEmail
 {
     /** @use HasFactory<\Database\Factories\UserFactory> */
-    use Billable, HasFactory, Notifiable;
+    use Billable, HasFactory, Notifiable, SoftDeletes;
 
     /**
      * The attributes that are mass assignable.
@@ -45,7 +46,9 @@ class User extends Authenticatable implements MustVerifyEmail
     {
         return [
             'email_verified_at' => 'datetime',
+            'is_locked' => 'boolean',
             'password' => 'hashed',
+            'deleted_at' => 'datetime',
         ];
     }
 
@@ -53,8 +56,11 @@ class User extends Authenticatable implements MustVerifyEmail
     {
         parent::boot();
 
-        // Delete all contacts when user is deleted
         static::deleting(function (User $user) {
+            if (! $user->isForceDeleting()) {
+                return;
+            }
+
             $user->contacts()->each(function (Contact $contact) {
                 $contact->delete();
             });
